feat: add ProductService class
- move CURD method from ProductController to ProductService
